+++ MemfaultPackages/bort-shared/src/main/java/com/memfault/bort/shared/Constants.kt
@@ -1,16 +1,5 @@
 package com.memfault.bort.shared
 
-const val INTENT_ACTION_BUG_REPORT_START = "com.memfault.intent.action.BUG_REPORT_START"
-
-const val INTENT_EXTRA_BUG_REPORT_MINIMAL_MODE_BOOL = "com.memfault.intent.extra.BUG_REPORT_MINIMAL_MODE_BOOL"
-const val INTENT_EXTRA_BUG_REPORT_REQUEST_ID = "com.memfault.intent.extra.BUG_REPORT_REQUEST_ID"
-const val BUG_REPORT_REQUEST_ID_MAXIMUM_LENGTH = 40
-const val INTENT_EXTRA_BUG_REPORT_REQUEST_REPLY_RECEIVER = "com.memfault.intent.extra.BUG_REPORT_REQUEST_REPLY_RECEIVER"
-const val INTENT_EXTRA_BUG_REPORT_REQUEST_TIMEOUT_MS = "com.memfault.intent.extra.BUG_REPORT_REQUEST_TIMEOUT_MS"
-
-const val INTENT_ACTION_BUG_REPORT_REQUEST_REPLY = "com.memfault.intent.action.BUG_REPORT_REQUEST_REPLY"
-const val INTENT_EXTRA_BUG_REPORT_REQUEST_STATUS = "com.memfault.intent.extra.BUG_REPORT_REQUEST_STATUS"
-
 const val DROPBOX_ENTRY_ADDED_RECEIVER_QUALIFIED_NAME = "com.memfault.bort.receivers.DropBoxEntryAddedReceiver"
 
 const val APPLICATION_ID_MEMFAULT_STRUCTUREDLOGD = "com.memfault.structuredlogd"